requirements.txt
Flask==1.1.2
gunicorn==20.0.4
numpy==1.19.5
scipy==1.6.0
google-cloud-firestore
firebase-admin
